fix: 모델 파이프라인 2차 수정 — 데이터 누수, 타입 불일치, 안전성
- cleaner.py: interpolation limit_direction='both'→'forward' (미래 데이터 누수 제거)
- retrain.py: evaluate 결과 안전 접근 (KeyError 방지)
- predict.py: 미등록 분야 ValueError, idx 안전 할당
- transformer.py: field_encoded .astype(int) 제거 (학습/서빙 타입 통일)
- ensemble.py: evaluate()를 가중 MAPE로 변경
- lstm_model.py: 패딩을 scaler 기반 중립값으로 변경

+++ edupulse/preprocessing/transformer.py
@@ -94,7 +94,7 @@ def add_lag_features(
         df["month_cos"] = encodings.apply(lambda t: t[1])
 
     if "field" in df.columns:
-        df["field_encoded"] = df["field"].apply(compute_field_encoding).astype(int)
+        df["field_encoded"] = df["field"].apply(compute_field_encoding)
 
     return df
